Better fix for trails
Rather than returning the end of the trail if a position for a segment
is beyond the end of the trail, use its "unit vector to previous
position" to extrapolate a position beyond the end.  This fixes a
continued problem which occurred after making a turn after a long
straightaway (which caused the previous position to be removed from the
list).